refactor: split synapse scoring.rs into boost, improvement, and discounting modules (#982)
Split the 877-line scoring.rs into a scoring/ module directory with
focused sub-modules:

- boost_functions.rs: source-type, target-type, activation-function boosts
- improvement.rs: core improvement calculation algorithm, candidate dedup
- discounting.rs: pessimism discounting, prediction calibration
- test_helpers.rs: test-only wrapper functions
- tests.rs: unit tests for improvement functions

All public API remains accessible through the same paths via re-exports
in scoring/mod.rs. No changes required in consuming code.
